splitting test files
