+++ src/dateAndTime.js
@@ -4,6 +4,7 @@
  * Getting an Array of Times + "AM" or "PM".
  * @param {number} startTime Starting hour [in minutes].
  * @param {number} minutesInterval every *n* minutes [default = 15].
+ * @returns {Array} Array of time.
  */
 const getTimes = (startTime = 0, minutesInterval = 15) => {
   const times = []; // time array
@@ -31,21 +32,51 @@ const getTimes = (startTime = 0, minutesInterval = 15) => {
   return times;
 };
 
+/**
+ * Returns Today's date.
+ * @returns {Date} Today's date.
+ */
 const today = () => {
   return new Date();
 };
 
+/**
+ * Returns Tomorrow's date
+ * @returns {Date} Tomorrow's date
+ */
 const tomorrow = () => {
   const day = new Date();
   day.setDate(day.getDate() + 1);
   return day;
 };
+/**
+ * Converts date into time stamp formate.
+ * @param {Date} date date
+ * @returns {Date} In time stamp formate.
+ */
 const toTimeStamp = (date) => {
   return new Date(date).getTime();
 };
+/**
+ * Converts date into UTC timezone.
+ * @param {Date} date date
+ * @returns {Date} In UTC timezone.
+ */
 const toUTC = (date) => {
   return new Date(date).toGMTString();
 };
+
+/**
+ * Converts Date into Human readable date string.
+ * @param {Date} date [default new Date()].
+ * @param {String} locales [default "en-Us"].
+ * @param {Object} options Formatting options.
+ * @param {String} options.weekday
+ * @param {String} options.year
+ * @param {String} options.month
+ * @param {String} options.day
+ * @returns {String} Human readable date.
+ */
 const humanFriendlyDate = (date = new Date(), locales = "en-US", options) => {
   const O = {
     weekday: "short",
@@ -56,15 +87,15 @@ const humanFriendlyDate = (date = new Date(), locales = "en-US", options) => {
   };
   return date.toLocaleDateString(locales, O);
 };
-const humanFriendlyTime = (date = new Date(), locales = "en-US", options) => {
-  const O = {
-    weekday: "short",
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-    ...options,
-  };
-  return date.toLocaleTimeString(locales, O);
+
+/**
+ * Converts Date into Human readable time string.
+ * @param {Date} date [default new Date()].
+ * @param {String} locales [default "en-Us"].
+ * @returns {String} Human readable date.
+ */
+const humanFriendlyTime = (date = new Date(), locales = "en-US") => {
+  return date.toLocaleTimeString(locales);
 };
 
 module.exports = {
