+++ src/objects.js
@@ -22,7 +22,25 @@ const toFormData = (obj) => {
  */
 const softClone = (obj) => JSON.parse(JSON.stringify(obj));
 
+/**
+ * Returns a unique array of objects based on a key.
+ * @param {array} array Array of objects.
+ * @param {string} key A unique property of each object [default = 'id'].
+ */
+const getUniqueObjs = (array, key = "id") => {
+  const ids = [];
+  const output = [];
+  array.forEach((ele) => {
+    if (!ids.includes(ele[key])) {
+      ids.push(ele[key]);
+      output.push(ele);
+    }
+  });
+  return output;
+};
+
 module.exports = {
   toFormData,
   softClone,
+  getUniqueObjs,
 };
