+++ src/objects.test.js
@@ -1,0 +1,14 @@
+import * as jsf from "./objects";
+
+describe("getUniqueObjs()", () => {
+  it("returns a unique array of objects", () => {
+    const arr = [
+      { id: 1, name: "1" },
+      { id: 1, name: "1" },
+      { id: 1, name: "1" },
+      { id: 1, name: "1" },
+    ];
+    const uniqArr = [{ id: 1, name: "1" }];
+    expect(jsf.getUniqueObjs(arr)).toEqual(uniqArr);
+  });
+});
